requirements.txt
# Core Framework Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# AI and ML Libraries
torch>=1.9.0
transformers>=4.20.0
openai>=0.27.0
anthropic>=0.3.0
langchain>=0.0.200
langchain-community>=0.0.10

# Natural Language Processing
spacy>=3.4.0
nltk>=3.7
textblob>=0.15.3
gensim>=4.2.0

# Knowledge Graph and Reasoning
rdflib>=6.2.0
networkx>=2.8.0
py2neo>=2021.2.0

# Web and API
requests>=2.28.0
aiohttp>=3.8.0
fastapi>=0.78.0
uvicorn>=0.17.0

# Data Processing
pydantic>=1.9.0
marshmallow>=3.17.0
jsonschema>=4.0.0

# Security and Safety
cryptography>=3.4.0
python-jose>=3.3.0
passlib>=1.7.4

# Monitoring and Logging
loguru>=0.6.0
prometheus-client>=0.14.0
structlog>=22.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=3.0.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.19.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Database
sqlalchemy>=1.4.0
redis>=4.0.0
motor>=2.5.0

# Configuration
python-dotenv>=0.19.0
pyyaml>=6.0
toml>=0.10.0

# Utilities
click>=8.0.0
rich>=12.0.0
tqdm>=4.64.0 